Update with latest versions
I recreated the sample application based on instructions from
https://create-react-app.dev/docs/getting-started/ as had been done
previously in this tutorial repository.

That updates the application and should replace the pull requests:

* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/100
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/138
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/140
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/141
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/142
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/143
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/144
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/146
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/147
* https://github.com/jenkins-docs/simple-node-js-react-npm-app/pull/150

https://react.dev/blog/2025/02/14/sunsetting-create-react-app announces
that as of Feb 2025, Create React App is deprecated for new apps.
Existing apps are recommended to migrate to a framework or to a build
tool like Vite.

https://react.dev/learn/creating-a-react-app#react-router-v7 notes that
React Router is the most popular routing library for React and can be
paired with Vite to create a full-stack React framework.  Since we're
using Vite for the new documentation site, it is probably best to use
React Router and Vite when we replace the sample application in this
repository.

Testing done:

* Ran the application with `npm start` and confirmed that it rendered
  as expected on port 3000
* Ran the `jenkins/scripts/test.sh` script and confirmed that tests pass
* Ran the `jenkins/scripts/deliver.sh` script and confirm that the site
  built and rendered as expected on port 3000